added cast for unsiged integer check

+++ klib/entry/entry.c
@@ -24,7 +24,7 @@ void __attribute__((__noreturn__, __naked__)) __reset_handler() {
     extern uint8_t __bss_end;
 
     // set the bss section to 0x00
-    for (uint32_t i = 0; i < (&__bss_end - &__bss_start); i++) {
+    for (uint32_t i = 0; i < ((uint32_t)(&__bss_end - &__bss_start)); i++) {
         ((volatile uint8_t*)(&__bss_start))[i] = 0x00;
     }
 
@@ -33,7 +33,7 @@ void __attribute__((__noreturn__, __naked__)) __reset_handler() {
     extern uint8_t __data_end;
 
     // copy rom to ram
-    for (uint32_t i = 0; i < (&__data_end - &__data_start); i++) {
+    for (uint32_t i = 0; i < ((uint32_t)(&__data_end - &__data_start)); i++) {
         ((volatile uint8_t*)(&__data_start))[i] = (&__data_init_start)[i];
     }
 
@@ -41,7 +41,7 @@ void __attribute__((__noreturn__, __naked__)) __reset_handler() {
     extern const entry_constructor __preinit_array_end;
 
     // excecute all the preinit constructors
-    for (uint32_t i = 0; i < (&__preinit_array_end - &__preinit_array_start); i++) {
+    for (uint32_t i = 0; i < ((uint32_t)(&__preinit_array_end - &__preinit_array_start)); i++) {
         // call the preinit calls
         (&__preinit_array_start)[i]();
     }
@@ -50,7 +50,7 @@ void __attribute__((__noreturn__, __naked__)) __reset_handler() {
     extern const entry_constructor __init_array_end;
 
     // excecute all the global constructors
-    for (uint32_t i = 0; i < (&__init_array_end - &__init_array_start); i++) {
+    for (uint32_t i = 0; i < ((uint32_t)(&__init_array_end - &__init_array_start)); i++) {
         // call every constructor we have
         (&__init_array_start)[i]();
     }
@@ -63,7 +63,7 @@ void __attribute__((__noreturn__, __naked__)) __reset_handler() {
 
     // run all the destructors. Should never be called but if it happens
     // it should work
-    for (uint32_t i = 0; i < (&__fini_array_end - &__fini_array_start); i++) {
+    for (uint32_t i = 0; i < ((uint32_t)(&__fini_array_end - &__fini_array_start)); i++) {
         // call every destructor we have
         (&__fini_array_start)[i]();
     }
